// packages/shared-logic/src/types.ts
export type UserStatus = 'active' | 'pending' | 'banned' | 'deletion_requested';
export type UserRole = 'user' | 'service_provider';

export interface AppUser {
  id: number;
  name: string;
  email: string;
  password?: string;
  avatar: string;
  status: UserStatus;
  role: UserRole;
  joinDate: string;
}
// ...
export type Theme = 'light' | 'dark' | 'system';